+++ src/App.jsx
@@ -4,7 +4,7 @@ import GameLoader from "./components/loaders/game-loader"
 function App() {
   return (
     <>
-    <GameLoader/>
+    <HomepageIndex/>
     </>
   )
 }